Add MummyNoise, wrapping several random values
MummyNoise seems quite a bit faster than some other recent kinds of
high-dimensional noise added to SquidLib, but currently has issues with
consistency between dimensionalities (this could very well be an issue
with the test in HashVisualizer). TabbyNoise remains slower than
MummyNoise, but now has a smoother appearance due to using
NumberTools.sway() instead of NumberTools.zigzag(),  and some other
small changes.

The project builds again; the most recent PR didn't compile in the
presence of GreasedRegion and CoordPackerZone, which needed
Zone.Skeleton to change getInternalBorder to return a Collection of
Coord (again) instead of a List, so overriding classes could return
Collections. getExternalBorder() just uses the code of getInternalBorder
inline, which allows it to store the result as a List.