Improve MersenneTwister instead of deprecating it
MersenneTwister now is based on an optimized version from a 2004 rewrite
instead of the badly-un-optimized multithreaded version we had
previously had in the source. The original code is BSD 3-clause licensed
and that is included in the source. ChaosRNG also has received changes
with the goal of removing dependence on java.security packages, since
they aren't available on GWT and may bloat the distributable size on
desktop with Java 9's jlink tool if security modules are needed.