Improvements to StringKit, MersenneTwister
StringKit.safeSubstring now is more robust when used in conjunction with
various indexOf methods. Since those often return -1 for "not found",
and that often happens at the end of a search through a String,
safeSubstring treats a negative endIndex as meaning "no end specified,
use the rest of the String", rather than trying to match at an
impossible spot and returning an empty String. StringKit also has
methods to get the index of a regular expression's first occurence in a
String, optionally after some start point.

MersenneTwister has reasonable toString(), equals(), and hashCode()
methods now, though there's no good reason to have a MersenneTwister in
a hashed collection as a key.

LZSPlus has some minor internal changes, and the test does some more
checks to make sure the encoding is robust enough to look completely
different if the keys are incorrect.