Grab bag of general improvements
CrossHash.equalityHelper() now correctly treats differently-sized input
arrays as not-equal, even if some of their contents are equal.
HashVisualizer has methods for hashing float positions in 1D on up, in a
way that can be optimized by running it on the GPU (see comments and
docs on floatHash() methods in that class). Various classes have javadoc
warnings fixed by adding better documentation.