Improve ThrustRNG quality, SeededNoise speed
ThrustRNG remains the fastest generator that doesn't have significant
statistical issues in SquidLib, and now it has had its quality improved
more. It passes TestU01's BigCrush test battery, which is one of the
most challenging tests for an RNG, on all tests for at least one seed
(testing on more seeds will take time, since each run of BigCrush takes
4 hours). The results of ThrustRNG calculations will change now, and
since Thrust's algorithm is used in WhirlingNoise, TabbyNoise,
MumyNoise, and CosmicNumbering, the results of those will change as well
(aesthetically, the changes seem either good or hard to notice,
depending on the dimensionality of noise).

Unrelated to ThrustRNG, testing those noise classes revealed an issue
with SeededNoise in 6D, where it was producing large amounts of garbage
when called continuously. This has been fixed, and the quality of 4D and
6D noise has been changed (hopefully for the better) in SeededNoise.

WhirlingNoise also had some small adjustments made to keep its results
between -1.0 and 1.0 for 3D noise.

Some small changes have been made to the tiling maps produced by
WorldMapGenerator; it may be hard to find defaults for noise frequency
that work for a variety of noise algorithms and implementations.