Fix for SquidMessageBox with IColoredString.
This is relevant for #16 . The implementation of appendMessage() for
String correctly added an item to the internal List of IColoredString
values, but appendMessage() for IColoredString did not. This should work
now.

I was considering adding a method to IMarkup to escape non-markup usage
of characters that might be mistaken for markup, but it doesn't seem
needed and would break anything that implements IMarkup. So, the code I
wrote is only present in comments in markup-related files.