Help SColor, SColorFactory. Don't use NumberUtils.
We already have NumberTools doing most of the same things as
NumberUtils, so we don't need to use an emulated layer to replace
libGDX's NumberUtils. SColor and SColorFactory have been updated to use
NumberTools instead of NumberUtils, fix some missing things, and
significantly improve docs. SColorFactory hasn't been relevant for a
while since we have SquidColorCenter, but one of the features it had was
the ability to look up a color by name. This functionality is already
present in libGDX's Colors class (at least, to look up a Color), so now
SColor populates the map from names to Color objects in Colors as it
constructs each named SColor (and only when an SColor has an assigned
name). This means SColorFactory.colorForName() is deprecated, since
using it requires populating a whole second map with String keys and
SColor values, and that effectively duplicates the map in Colors but
only holds SColor values. Having Colors be populated with all named
SColors means that hand-written markup using GDX's color markup style
can use named colors instead of hex color codes.