Random fill methods in ArrayTools
These use a variant on a new algorithm that may replace ThrustRNG's
current algorithm at some point; since this new variant supports a
fairly large number of different streams, it makes sense to use it for
multidimensional random value generation, which is what ArrayTools can
do now. This was meant to be convenient, so there are methods to bound
the generated values (as int, float, or double) and to choose random
char values from a String or char array.