New approach for block glyphs; TCF temp change
Most of the distance field fonts now include the full block image at
Unicode codepoint 0. That block was previously a separate Texture, and
that has performance downsides when mostly rendering text. By keeping an
entire text-based line as potentially one String with different colors,
we also open up possibilities for both optimizing our drawing a little
better, and for using markup on long Strings (as I'd like to have for
#16 )Though the fonts should keep the changes in this commit for later
(it gave me a chance to improve some unrelated spacing issues in some
fonts), TextCellFactory will have its changes in this commit immediately
reverted, since not all fonts have a block at codepoint 0, including
probably all user-created fonts. The solution is probably to use a
subclass of TextCellFactory for fonts we know have been made to use the
block in the font, and default to the normal separate texture for all
other fonts.