Revert changes to TCF. Remove some unused assets.
The icon atlas now no longer includes GUI controls because most games
would probably want their own kind of GUI anyway, and including them
significantly increased the size of the atlas and assets.zip. The
"SciFi" font Galaxsea-Starlight-Mono-v3.1 has been removed from the
assets and DefaultResources, because it was just not clear enough with
the distance field effect in use and it's just frustrating to play a
text-based game when you can't read the font.

TextCellFactory goes back to how it was with this commit, using a block
for solid cells. It performs well enough either way, so the markup usage
would be the deciding factor.