Early work to support bold/italic for fonts
TextFamily extends TextCellFactory with support for loading groups of
fonts from atlases. It currently has two choices of valid fonts
(Iosevka, and Iosevka Slab). TextCellFactory has a method to change the
current face, which does nothing in that class but can be used in
TextFamily to go from the current style to bold, italic, bold+italic, or
back to regular. This needs more support for changing the style
mid-render, since this is capable of it now.

This commit completes removal of Galaxsea Starlight Mono.