Support more dice notation features
Some of these were possible already, such as by calling bestOf() with a
String containing dice notation, but couldn't be used with just
roll(String). Now you can get the best of or worst of exploding dice as
well as regular dice, and can get ranges with dice results or other
ranges as the upper bound (this can be useful to curve distributions).
Thanks again to Rakaneth for mentioning this; I hadn't made the
connection that bestOf() could use exploding dice, but roll() couldn't
directly.