Italics and bold work! GDXMarkup can use them.
Though it isn't present in this commit, extra code may be able to add
bold and italic chars via a TextFamily directly. Bold and italic are
encoded in the upper two bits of a char, which doesn't affect any of the
"permissible chars" as specified in TextCellFactory and StringKit for
chars that are known to behave correctly. This may need to be changed if
people are using chars in the upper 3/4 of the Unicode BMP.