Slight change; allow full Unicode for TCF (again)
TextFamily now is the only part of rendering with a restricted range of
Unicode. TextCellFactory renders all of the Unicode BMP again, and
should have a little better performance when drawing. This screenshot
has italic, bold, and bold italic fonts on display on the bottom row of
the message box: https://i.imgur.com/P9cY90h.png .