Adjust new blend() method, GWT fix in NumberTools
The blend() method is in SparseLayers and SquidPanel, and the new
behavior it has makes some useful PanelEffect implementations look
nicer. This uses the new SColor.lerpFloatColorsBlended() method, which
is very similar to the previous float-lerping code but treats alpha
differently. Docs have been added in some places. NumberTools had a
method added to the JDK version, but not the GWT emulated version; this
has been fixed.