Add ThrustAltRNG, fix ArrayTools.randomFill()
ThrustAltRNG is a variant on Thrust's general algorithm, with all the
numbers specifically adjusted to avoid statistical quality issues,
testing with gjrand and PractRand. It has a small speed loss compared to
ThrustRNG, but can get through over 1TB of tests with PractRand instead
of the 32GB that kills ThrustRNG in testing (it starts to suffer
problems at 2TB). ArrayTools used an earlier version of ThrustAltRNG's
algorithm, but a mistake in porting from C (which uses `>>` with
unsigned ints to mean what `>>>` does in Java) caused various issues in
the less-than-random choices it made before. This has been fixed.