Downgrade to known-working GDX 1.9.6
It didn't take long for a bug in libGDX 1.9.7 to occur and prevent one
of my game projects from running at all on GWT. The bug is related to
TextureAtlases in the specific case, but more broadly affects GWT code
that involves the libGDX Array class (and probably other GDX
collections). Until it's fixed (not holding my breath...), we're
downgrading SquidLib to 1.9.6 and encourage other libraries to not
update to 1.9.7 either, since it's very hard to mix libGDX versions
effectively and 1.9.7 looks pretty bad so far.