Update ThrustAltRNG with better skip, determine
The new algorithm seems to have near-identical performance on next() and
nextLong(), but better quality in long-running cases, and because of the
structure being different, skip() and determine() are much simpler.
WhirlingNoise uses a slightly-slower variant on ThrustAltRNG.determine()
now to choose gradient vectors, and it seems to significantly reduce the
occurence of long lines with the same or similar gradient vector (which
appear as solid white or black lines in the 2D noise).