Fix SquidPanel positioning bug. Close #186
SquidPanel previously didn't take into account x and y changes applied
to it as a scene2d Actor; now it does. This shouldn't affect much
existing code, but it may affect some code that changed the position of
a SquidPanel or SquidLayers, especially if AnimatedActor values are
involved. A test for this is in PositionTest. ThrustAltRNG is having
some experiments done.