Update SquidPanel with Rakaneth's proposal
In #188 , Rakaneth mentioned that SquidMessageBox already has code to
generate border cells, which I had completely forgotten about. This
commit makes that code public and moves it to SquidPanel, but changes it
from returning a new 2D char array on each call to actually placing the
chars directly in the border cells. It leaves the inner cells alone,
unlike the old method in SquidMessageBox, so methods have been added to
ArrayTools for filling a subsection of a 2D array conveniently (this is
now used by SquidMessageBox to clear the non-border cells). This also
fixes a weird bug in EverythingDemo, where the health text wasn't
appearing where the border should have been overwritten by a number in
red; now the text appears as it should. An upcoming commit may also add
border-drawing functionality to SparseLayers. Thanks, Rakaneth !