Trying to fix some GWT/desktop discrepancies
In my project ColorGuard, I have a GWT and a desktop module, and the
generated world is completely different despite the seed being the same.
I think this is related to hashing quirks on GWT or something in
FakeLanguageGen, but likely both. Calling hashCode() on one
FakeLanguageGen during debugging yielded an absurdly large number,
bigger than an int can hold and shown as a double, so I'm moving
calculations that previously relied on hashCode to a 64-bit hash method.
This will change any mixes of languages that used the two-argument mix()
method, and I don't know if it affects other mixes. It shouldn't affect
unmixed languages.