SquidPanel: can put String with float color
This has been added to the new code from Rakaneth's PR, which I should
have given access to this method earlier but didn't because I forgot to
add it as a counterpart to the put() with a Color arg. There are also
options to add IColoredString captions now, and this is used in
EverythingDemo. These IColoredString-caption versions are named
differently because a null parameter could be either a String or an
IColoredString if they were both called putBorders().