Convenience for scaled noise, compat for OrderedMap
Noise.Scaled2D, 3D, 4D, and 6D now have constructors that set all axes
to the same scale amount with one parameter; this is just a minor
convenience but it does cut down on repetition. OrderedMap now has a
static method to instantiate an OrderedMap value given alternating keys
and values; this is almost identical to an existing method in Maker but
has more intuitive behavior when Float keys or values are involved,
since Maker also takes a float for the load factor and that can confuse
javac. This also gives some compatibility for TabLabTools, another
project I'm working on that I'd like to be able to use SquidLib more
easily, and the name makeMap() establishes a convention for other
key-value-pair instantiation methods.