Can use NatLangCipher on part of a String now
This uses NaturalLanguageCipher.cipherMarkup(), which uses markup that
is similar to and compatible with GDXMarkup. Adding ? in square brackets
toggles translation on (or off, if it's already translating), and it
keeps the non-word characters the same so it can translate a block of
text. This is intended to be used for cases where a character would be
referring to some named character and the name should be unique to the
language used, if a character suddenly breaks out into another language
due to surprise, or all sorts of other possibilities.