Fix #190 , add more Dice tests
The issue was indeed related to the simple regex parsing for dice
notation, and that was a one-char fix (changing `(-?\\d+)` to
`(-?\\d+)?`, making the first digit optional). But, there was also
another bug (unrelated to the original issue); too much whitespace could
prevent operators from being parsed correctly. There's a test for this,
but I had not been running the tests during a frequent build cycle... or
maybe a failed test wasn't stopping the build. It probably was the
latest change, making the whole dice string optional so any part can be
omitted, that allowed spaces to be considered a viable part of a dice
string by the parser, and so made `  *  ` parse as the default "add 0",
then the asterisk was parsed as "multiply" but promptly overridden by
another "add 0". Spaces are now fully ignored between sections.