Change RNG defaults to use ThrustAltRNG
The big change here is that where LightRNG was the default before for
pretty much all RNG-related stuff, ThrustAltRNG is now used instead.
ThrustAltRNG has about the same features (except the stuff that's
duplicated in RNG, like a nextDouble() method), including being a
StatefulRandomness and a SkippingRandomness. SkippingRandomness is new
in this commit, and it is an interface that marks RandomnessSources that
you can skip() forwards and backwards in. Some procedural stuff that
relied on a seed but didn't specify a RandomnessSource will change
generation behavior. ThrustAltRNG is a fair bit faster than LightRNG,
and a benchmark against best-of-class shuffling algorithms, where
ThrustAltRNG did very well, is what made me decide to switch.

Some docs still need to be updated.