SColor.toEditedFloat() and related methods added
The shader used by ShaderDemo, which is present in DefaultResources, now
allows toggling the outline effect as off or on by setting the smoothing
multiplier to its current value * -1. The shader looks a lot better if
the interior is a lighter color, which was a good reason to introduce
toEditedFloat(), which can adjust the HSV components of a color as it
converts it to a float.