MerlinNoise now implements Noise interfaces
There's also some possible optimizations used in the point hashing for
MerlinNoise, but it's hard to gauge if they help because about 85% of my
CPU has been in use on various RNG tests for a few days. The new way
does seem to require less operations and method calls than the old way,
at least.