Subclasses of RNG should copy() with that class
Not sure why this wasn't the case before, and I can't think of a case
where this would break code. It is a lot more convenient to make a copy
of a StatefulRNG with `StatefulRNG other = mySRNG.copy()` than the older
alternative `StatefulRNG other = new
StatefulRNG(mySRNG.getRandomness().copy)`. This also works for other
classes where that alternative wouldn't be enough.