Major changes to EditRNG, fixes in RandomBias
EditRNG had some very strangely-shaped and unintuitive distributions
before, and was very slow due to heavy usage of Math.pow() and
trigonometry functions. It now uses a much conceptually simpler
technique (choose each time between a mountain-shaped distribution and a
valley-shaped one with the same average and range, using the centrality
value to decide). This doesn't yield nice smooth curves in the
distribution, instead producing simpler shapes like the triangular
distribution, but it is accurate and fast (though it generates 3 longs
per random value now, the RandomnessSources that it can use as a
StatefulRandomness are all very-highly-optimized).

RandomBias could sometimes return 1.0 from nextDouble() when using its
"bathtub" distribution, which made places that assumed it used the same
rules as RNG (which never returns 1.0 from nextDouble() ) and used
nextDouble() as part of an array indexing operation sometimes throw
exceptions at random when it went out of bounds. This is fixed.