Serious optimizing on DijkstraMap
DijkstraBenchmark has been updated to use the correct JMH techniques.
Finding smaller paths on 192x192 maps, DijkstraMap has gone from taking
1500+ ms to find 9908 length-9-or-less paths, to taking rougly 380 ms.
This uses a nice optimization for when both a starting point and a limit
(used by partialScan) are specified, by avoiding any preparation work on
parts of the map that will never be possible to scan with that limit;
this improves findPath() when scanLimit is used. Several cases also
benefit from not allocating int arrays internally in
appendDirToShuffle(), which is called often.