Add WorldMapGenerator.EllipticalMap, biome change
Because EllipticalMap can produce cells on the map that are empty space
(the projection is an ellipse, and the map is a rectangle), a new biome
type is needed to allow empty space. If you use WorldMapGenerator and a
custom biome generator, take a look at the changes to the biome mappers
in WorldMapGenerator and how their usage has slightly changed in
DetailedWorldMapDemo and DetailedWorldMapWriter. Preview in
https://squidpony.github.io/SquidLib/EllipseWorld.png .