FantasyPoliticalMapper for biome-sensitive nations
Elves need to be in forests! Snakefolk need to live in warm places!
Although human nations don't really care what the climate is if it can
be colonized (pretty much all of Earth has been claimed), fantasy
races/species often do care where they live or can't live outside some
area. This new class lets the colonization habits of such species work
by creating Faction objects that specify what biomes a group is willing
or unwilling to colonize, as well as preferred height, moisture level,
and heat as factors. It works well with EllipticalMap, handling wrapping
in its case as well as in TilingMap's case while keeping area on the
globe equal for equal-sized nations when using those two styles. It will
act somewhat weird with SphereMap and SphereMapAlt, with lots of nations
likely to start near the poles due to distortion, despite a much smaller
area of land to colonize; the same is true for PoliticalMapper.