Some progress on allowing political maps to zoom in
They can zoom in just fine, but they can't zoom out correctly yet. This
is related to the issues with rivers and zooming from earlier. On the
plus side, the edge of nations tends to be nice and ragged, like it
matches some invisible river or mountain range, which mimics many early
nation borders. This is thanks to the new GreasedRegion.fray() method,
which quasi-randomly removes cells from the surface of a GreasedRegion.