Add FakeLanguageGen.MAORI
This could be useful for some of the world-spanning game concepts I've
tinkered with, since Maori is probably the most distinctive language
from the Oceania area to English-speakers. I think the imitation should
sound fairly good; it uses only one letter with a diacritic (a macron)
even though actual Maori can use a macron on all vowels, just very
rarely for anything but 'a'. That should make it fairly well-supported
by fonts.