Bunch of missing features in SparseLayers now work
Only some methods for effects allowed a delay and/or postRunnable
before; now all of them have overloads that optionally allow setting
them (delay can be 0 and postRunnable can be null if either doesn't
apply for that effect). The reverse of this is now added to tint();
before it always took a delay and Runnable, and now those can be omitted
to treat them as 0f and null (the previous methods are all still there).