Add UnorderedMap, improve default load factors
UnorderedMap is just like HashMap but allows a custom IHasher to be used
and may be a little faster; there's a subclass coming soon that will
take advantage of the IHasher for convenience. OrderedSet and
UnorderedSet used a load factor of .375 by default, which turns out,
despite using more memory, to worsen performance in testing relative to
a load factor of 0.5 or the very similar (but more memory-friendly) 0.75
. So, the load factor is 0.75 now by default, and it can still be set
manually if you find it matters.