Issue found in ThrustAltRNG; probably minor
ThrustAltRNG can't produce some results with nextLong(), and produces
some values more than others, which seems good for actual random
results, but means some potentially possible results are actually
impossible. The actual impact may not be very significant, since
ThrustAltRNG passes every empirical quality test I can throw at it, but
it does prevent the existence of an inverse to its algorithm, while
LightRNG does have an inverse (which would take the output of
LightRNG.nextLong() and produce the state of the LightRNG). Both
ThrustAltRNG and LightRNG allow direct access to their state anyway, so
that doesn't matter much either. I'm working on a possible replacement
to ThrustAltRNG, if I can get it faster than LightRNG. Depending on the
results there, LightRNG may become the default again, or the successor
may be the default if it has very strong testing results.

This commit changes some of the point hashing to use LightRNG's
algorithm instead of ThrustAltRNG, and is able to perform some extra
optimizations for the hashes that get 8 bits or less.