Adjustments to Glyph, use Glyph more to fix #198
In fairly limited testing, the changes to TextCellFactory.Glyph seem to
line up the tinted foreground and background chars exactly with their
un-tinted counterparts in SquidPanel, which should address #198 . I
don't know yet if all the effects work correctly in SquidPanel, since
several things needed changing (from a generic Actor in most places to a
Glyph specifically, though this probably isn't usually needed any more,
and Glyph is a kind of Actor). Glyph previously stored a Color by virtue
of being an Actor, but didn't use it; now it does use it and this makes
it more compatible with other Actors and Actions. I'm also not 100% sure
ImageSquidPanel works fully right now, though I think it should. On the
plus side, there's no changes necessary in SparseDemo (so the demo
SquidSetup makes is still fine), and the only changes I made to
BasicDemo2 were to have a tint effect to test and to add some extra
niceties; nothing broke in the used API.