The randomized methods in NumberTools work again
While some of the methods would run fine on desktop, they weren't the
same as the super-sourced versions on GWT, since those still used
ThrustAltRNG's algorithm and the desktop NumberTools used LightRNG.
Making the problem worse, a bug in porting swayRandomized to use
LightRNG meant that it didn't produce a continuous curve. Now all of
these use ThrustAltRNG's algorithm, because it is faster on this
generally-CPU-intensive code, because it is cleaner to write with less
temporary variables, and because ensuring unique long output is not
exactly a priority when these are all being converted to floats or
doubles anyway.