Fixes uncommon issue with addAt() in OrderedSet
The issue boiled down to IntVLA.moveToFirst() (and also moveToLast())
being used as if it expected a value, when in fact it expected an index.
Since the only code in SquidLib that used moveToFirst() was collections
code to implement addAndMoveToFirst(), getAndMoveToFirst(), or some
similar method, it makes sense to change IntVLA to match the
expectations, even though they were incorrect before. That said, there's
no reason to use addAndMoveToFirst() because we have addAt(0), and so
addAt(0) now no longer delegates to addAndMoveToFirst(), instead using
the better-optimized existing code. This means the existing code that
would have encountered the bug in unrelated code now won't have any
dealing with the old formerly-buggy code, and code that used the
fooAndMoveToFirst methods will actually work, hopefully reliably now.