Make LOS brush mode work as documented
Checking the different LOS options in LOSComparisonTest, brush had some
weird issues where it would fail to draw a line even if DDA or Bresenham
would succeed. Now brush mode will consider a target point reachable if
any line can be drawn from a point offset from the start to a point with
the same offset from the target. Elias mode of LOS also shouldn't
duplicate a point in the middle of the List it returns.

LOSComparisonTest now shows if a line of sight connected with a target
(shows `$`), failed to connect to a target (shows `?`), is a connecting
line over a floor (shows `*`), or the normal `@` for start, `.` for
floor, or `#` for wall (even if a wall has a line end inside it).