VDCQRNG has a nice new subrandom sequence
VanDerCorputQRNG.altDetermine() seems to sometimes get better results on
minimum distance between points than the Halton sequence does, for no
clear reason. The best bases found are 0xDE4DBEEF and 0x1337D00D, which
seems like a joke but I don't know why they do so much better. Using
0x1337DEED instead of 0x1337D00D makes it like a Halton sequence for its
minimum distance (it does compute slightly faster). It's certainly
unusual, but it's hard to find any improvement to Sobol or Halton
sequences for this type of measurement, so it's definitely a win, but a
puzzling one.