Avoid problematic Sobol in GreasedRegion
The GreasedRegion.randomSeparated() method doesn't reliably fill any
specific amount of the limit it's supposed to fill, and there are three
different alternatives available, so I've removed usage within SquidLib
of randomSeparated() and other methods in GreasedRegion that use a Sobol
sequence (which is a poor fit for the task) to get the separated points.
The docs have been substantially updated in GreasedRegion.