Potentially breaking: SquidInput allows null mouse
If you didn't specify a SquidMouse before, SquidInput would still make
one that didn't do anything, and there was no way to specify that a game
or component of a game was purely keyboard-driven. Now SquidInput will
have a null SquidMouse if none is given, so code that called getMouse()
after none was given will now throw a NullPointerException (that
shouldn't have been allowed in the first place, but was used and had no
effect in ImageDemo here in SquidLib).

Also, nice additions to FakeLanguageGen so you can get the
human-readable name of a language with getName(), which also works on
mixes of languages. The Maven plugins have been updated to the current
version, which only needed small changes to anything.