SquidInput loads file "keymap.preferences"
It does this only if local files can be read and if that file exists,
but when it does it is used to fill the key remappings in an automated
way that doesn't require user input (so it can be used when the app is
starting up). An application should be made soon, probably in the
tommyettinger/SquidLib-Demos repo, that will be a simple tool to produce
remapping files for different keyboards and preferred settings. These
files are meant to be usable across games, so a main project using
SquidLib (like SquidPony/Epigon ) and a gamejam project (like
SquidPony/DD-7dRL2018 ) can share these files and not require
reconfiguration.