Add KeyRemapTool with instructions, a preset
You can test the auto-remap behavior of SquidInput by copying
keymaps/vi-keys/keymap.preferences into the root of the SquidLib repo,
then running SparseDemo as a test. It should respond to vi-keys (hjkl
for orthogonal directions) if the keymap was correctly loaded. If you
later move keymap.preferences out of that directory, SparseDemo won't
respond to vi-keys any more (it was edited in this version to test this
behavior, and will default to allowing vi-keys later).

KeyRemapTool is from tommyettinger/SquidLib-Demos, and was minified
using a somewhat-challenging-to-use tool that's also kept in that repo,
JarShrink. It's possible that JarShrink will work out of the box for
other people, and I tried to enable that, but I really don't have a good
way to test it on another development machine. It does reduce the size
of runnable JARs substantially without requiring the dark arts of
ProGuard config.