Fix weird changing-sea-level bug in WMG
This changes a lot of world map stuff, because the cause of the bug was
an adjustment that increased the ratio of land to water, based on the
min and max visible height. So land amounts needed to be re-adjusted;
now land is almost as common as water (it was initially less common but
the adjustment had overcompensated) to allow for both "blue planets"
like Earth with lots of water as well as drier worlds. Weirdly,
ClassicNoise is a lot faster here than WhirlingNoise or SeededNoise in
3D, plus it seems to produce rather good quality, especially when using
WhirlingNoise for 4D as SpaceViewMap does. Rotating space view using
ClassicNoise: https://i.imgur.com/XL0qnzR.gifv