Fix some issues with hashes in gradCoord methods
Their hashes weren't getting all possible indices for the corresponding
gradient vector arrays, so the noise was overly gray (too close to the
mid-range) and the adjustments to keep the result in-range was
incorrect. Now ClassicNoise and 3D WhirlingNoise have more results near
the high and low ends of the range (white and black) and generally less
artifacts.

There's also the start of some work on an improved quasi-random sequence
based on Morton decoding a Weyl sequence or similar not-very-random
number.